+++ Sources/CorePayments/UIApplication+URLOpener.swift
@@ -7,6 +7,7 @@ public protocol URLOpener {
         universalLinksOnly: Bool,
         completionHandler completion: ((Bool) -> Void)?
     )
+    func isPayPalAppInstalled() -> Bool
 }
 
 public extension URLOpener {
@@ -17,6 +18,14 @@ public extension URLOpener {
 }
 
 extension UIApplication: URLOpener {
+
+    @available(*, deprecated, message: "No longer used for app switch eligibility. Will be removed in the next major release.")
+    public func isPayPalAppInstalled() -> Bool {
+        guard let payPalURL = URL(string: "paypal://") else {
+            return false
+        }
+        return canOpenURL(payPalURL)
+    }
 
     public func open(
         _ url: URL,